Split worktree removal tests by responsibility
The worktree removal suite combined cross-scene coordination, failed-removal recovery, confirmation, and reconciliation in one 3,900-line source file. Those behaviors already have separate seams, but the shared file made focused ownership and review difficult.

Keep all 63 tests in the original serialized suite while moving each seam to a responsibility-named extension. Shared fixture access expands only within the test target; production code, test identities, and scheduling remain unchanged.
